requirements.txt
uvicorn
fastapi
bcrypt
pydantic
starlette
deepl
bcrypt
python-dotenv